Add output section for relocations to linker script
Add an output section to the linker script to gather all of the
relocation sections into.  Without this we appear to trigger a bug in
the linker where the relocation sections are placed as orphans, which
results in an incorrect LMA, even though the relocation sections are
then discarded from the link.

What happens is we have a snippet of linker script like this:

  .text ALIGN (0x1000) : { *(.text) }
  .data ALIGN (0x1000) : { *(.data) }

and the .data is non-zero sized, while the .text is zero sized.  The
linker script contains no specific placement for .rela.* sections, and
so, tese are placed as orphans after .text, which does not change the
VMA but is (apparently) changing the LMA.

So, as the .text is zero-sized, the .data is placed at the same
aligned address as the .text, however, the placment of the orphaned
.rela.* sections has increased the LMA for .data.

To add more confusion, this is a final link, so the .rela.* sections
are subsequently discarded from the link, making the .rela.* output
section zero-sized.

It feels like there's a linker bug in there somewhere, but as an
immediate fix, simply creating an output section for the .rela.* input
sections will resolve the problem.

Files Changed:

	* p/link.ld: Add .rela.dyn output section.